+++ src/lib/mrvUI/mrvMonitor_macOS.cpp
@@ -29,5 +29,11 @@ namespace mrv
 
             return out;
         }
+        
+        float getMaxNits(int screen = -1, const bool silent = true)
+        {
+            float out = 1000.F;
+            return out;
+        }
     } // namespace monitor
 } // namespace mrv