+++ tlRender/lib/tlTimelineVk/RenderPrivate.h
@@ -132,6 +132,8 @@ namespace tl
             int32_t frameIndex; // must be an int32_t not an uint32_t.
 
             bool hdrMonitorFound = false;
+            float monitorMinNits = 0.005F;
+            float monitorMaxNits = 1000.F;
 
             math::Size2i renderSize;
             timeline::OCIOOptions ocioOptions;